requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-amazon-sp-api==5.0.1
gunicorn==21.2.0
pathlib2==2.3.7